+++ lnd_messages/close_channel_request.js
@@ -10,7 +10,7 @@ const hashAsTxId = hash => hash.slice().reverse().toString('hex');
     delivery_address: <Request Cooperative Close Address String>
     force: <Force Close Channel Bool>
     max_fee_per_vbyte: <Max Fee Tokens Per VByte String>
-    sat_per_byte: <Chain Fee Tokens Per Virtual Byte String>
+    sat_per_vbyte: <Chain Fee Tokens Per Virtual Byte String>
     target_conf: <Target Confirm Within N Blocks Number>
   }
 
@@ -31,7 +31,7 @@ module.exports = args => {
     is_force_close: args.force || undefined,
     max_tokens_per_vbyte: Number(args.max_fee_per_vbyte) || undefined,
     target_confirmations: args.target_conf || undefined,
-    tokens_per_vbyte: Number(args.sat_per_byte) || undefined,
+    tokens_per_vbyte: Number(args.sat_per_vbyte) || undefined,
     transaction_id: hashAsTxId(args.channel_point.funding_txid_bytes),
     transaction_vout: args.channel_point.output_index,
   };