+++ test/lnd_methods/onchain/test_send_to_chain_addresses.js
@@ -21,6 +21,20 @@ const makeArgs = overrides => {
   return args;
 };
 
+const makeExpected = overrides => {
+  const args = {
+    confirmation_count: 0,
+    id: Buffer.alloc(32).toString('hex'),
+    is_confirmed: false,
+    is_outgoing: true,
+    tokens: 1,
+  };
+
+  Object.keys(overrides).forEach(k => args[k] = overrides[k]);
+
+  return args;
+};
+
 const tests = [
   {
     args: makeArgs({lnd: undefined}),
@@ -75,35 +89,17 @@ const tests = [
   {
     args: makeArgs({log: console.log, wss: [{clients: [null]}]}),
     description: 'Send coins with null wss client',
-    expected: {
-      confirmation_count: 0,
-      id: Buffer.alloc(32).toString('hex'),
-      is_confirmed: false,
-      is_outgoing: true,
-      tokens: 1,
-    },
+    expected: makeExpected({}),
   },
   {
     args: makeArgs({log: console.log, wss: [{clients: [{}]}]}),
     description: 'Send coins with no ready state',
-    expected: {
-      confirmation_count: 0,
-      id: Buffer.alloc(32).toString('hex'),
-      is_confirmed: false,
-      is_outgoing: true,
-      tokens: 1,
-    },
+    expected: makeExpected({}),
   },
   {
     args: makeArgs({log: () => {}, wss: [{clients: [{readyState: 1}]}]}),
     description: 'Send coins with no send method',
-    expected: {
-      confirmation_count: 0,
-      id: Buffer.alloc(32).toString('hex'),
-      is_confirmed: false,
-      is_outgoing: true,
-      tokens: 1,
-    },
+    expected: makeExpected({}),
   },
   {
     args: makeArgs({
@@ -111,35 +107,22 @@ const tests = [
       wss: [{clients: [{readyState: 1, send: () => {}}]}],
     }),
     description: 'Send coins with broadcast',
-    expected: {
-      confirmation_count: 0,
-      id: Buffer.alloc(32).toString('hex'),
-      is_confirmed: false,
-      is_outgoing: true,
-      tokens: 1,
-    },
+    expected: makeExpected({}),
   },
   {
     args: makeArgs({}),
     description: 'Send coins',
-    expected: {
-      confirmation_count: 0,
-      id: Buffer.alloc(32).toString('hex'),
-      is_confirmed: false,
-      is_outgoing: true,
-      tokens: 1,
-    },
+    expected: makeExpected({}),
+  },
+  {
+    args: makeArgs({fee_tokens_per_vbyte: 1}),
+    description: 'Send coins with fee rate',
+    expected: makeExpected({}),
   },
   {
     args: makeArgs({utxo_selection: 'random'}),
     description: 'Send coins with coin selection',
-    expected: {
-      confirmation_count: 0,
-      id: Buffer.alloc(32).toString('hex'),
-      is_confirmed: false,
-      is_outgoing: true,
-      tokens: 1,
-    },
+    expected: makeExpected({}),
   },
 ];